backend/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.10
mutagen==1.47.0
pycryptodomex==3.20.0
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.3
websockets==13.1
yt-dlp==2024.9.27
